+++ Changelog.md
@@ -5,6 +5,7 @@ Breaking Changes
 ----------------
 - Extend `Tracer::span` with `StartOptions`.
 - `FileTracer::new` returns a `Tracer` and not a `FileTracer`.
+- Rename `ImplWrapper` to `ImplContextBox`.
 
 Features
 --------